refactor: customer info tab - make integrations a separate component
