+++ src/components/InfoRow/index.tsx
@@ -1,4 +1,6 @@
-const InfoRow = ({ children }: { children: React.ReactNode }) => (
+import { FC } from 'react'
+
+const InfoRow: FC<{ children: React.ReactNode }> = ({ children }) => (
   <div className="mb-2 flex gap-4 first-child:w-50 first-child:shrink-0">{children}</div>
 )
